requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
bidict==0.22.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
multidict==6.0.2
pycparser==2.21
pynput==1.7.6
pypiwin32==223
python-engineio==4.3.2
python-engineio-client==1.1
python-socketio==5.6.0
python-socketio-client==1.1
pywebostv==0.8.7
pywin32==304
requests==2.27.1
six==1.16.0
urllib3==1.26.9
ws4py==0.5.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
